+++ src/lib/RbtSetupPMFSF.cxx
@@ -163,7 +163,7 @@ RbtPMFType RbtSetupPMFSF::GetPMFfor_rS(RbtAtomPtr anAtom)
 void RbtSetupPMFSF::SetupReceptorPMFTypes(void) 
 {
 	//cout << _CT << " receptor size " << theReceptorList.size() << endl;
-	for (long i=0; i<theReceptorList.size(); i++) {
+	for (size_t i=0; i<theReceptorList.size(); i++) {
 		switch (theReceptorList[i]->GetAtomicNo()) {
 			case 1: theReceptorList[i]->SetPMFType(HH); 
 				break;
@@ -366,7 +366,7 @@ RbtPMFType RbtSetupPMFSF::GetPMFfor_lS(RbtAtomPtr anAtom)
 
 void RbtSetupPMFSF::SetupLigandPMFTypes(void)
 {
-	for (int i=0; i<theLigandList.size(); i++) {
+	for (size_t i=0; i<theLigandList.size(); i++) {
 		switch (theLigandList[i]->GetAtomicNo()) {
 			case 1: theLigandList[i]->SetPMFType(HL);  break;
 			case 6:		// carbon
